Move compatibility tests for recursive normalizers and reduce number of chunks for test_write_large_df_in_chunks test
